QueryDSL plugin install

+++ src/main/java/arc/ua/arcshop/dto/GoodsDTO.java
@@ -10,10 +10,10 @@ public class GoodsDTO {
     private String name;
     private String type;
     private String manufacturer;
-    private Integer price;
+    private Double price;
     private Integer count;
 
-    public GoodsDTO(String name,String type, String manufacturer, Integer price, Integer count) {
+    public GoodsDTO(String name,String type, String manufacturer, Double price, Integer count) {
         this.name = name;
         this.type = type;
         this.manufacturer = manufacturer;
@@ -21,7 +21,7 @@ public class GoodsDTO {
         this.count = count;
     }
 
-    public static GoodsDTO of(String name, String type, String manufacturer, Integer price, Integer count) {
+    public static GoodsDTO of(String name, String type, String manufacturer, Double price, Integer count) {
         return new GoodsDTO(name, type, manufacturer, price, count);
     }
 }
